💄 BoardList, BoardEnterPage storybook 수정
