+++ src/stories/components/BoardList.stories.ts
@@ -3,9 +3,22 @@ import { Meta, StoryObj } from '@storybook/react';
 
 const meta: Meta<typeof BoardList> = {
   component: BoardList,
+  argTypes: {
+    width: {
+      control: 'text',
+    },
+    height: {
+      control: 'text',
+    },
+  },
 };
 
 export default meta;
 type Story = StoryObj<typeof BoardList>;
 
-export const Deafult: Story = {};
+export const Deafult: Story = {
+  args: {
+    width: '428px',
+    height: '40px',
+  },
+};
