--- src/stories/pages/BoardEnterPage/BoardEnterPage.stories.ts
@@ -1,11 +1,0 @@
-import BoardEnterPage from '@/pages/BoardEnterPage';
-import { Meta, StoryObj } from '@storybook/react';
-
-const meta: Meta<typeof BoardEnterPage> = {
-  component: BoardEnterPage,
-};
-
-export default meta;
-type Story = StoryObj<typeof BoardEnterPage>;
-
-export const Deafult: Story = {};
